chore: add changes to server/** to test GitHub workflows
This commit adds a console.log statement to the code in the server/** directory
to check whether the GitHub workflows are working properly. The statement will
print a message to the console when the workflows are triggered by changes to
the server directory. This change is only a development tool and does not affect
the functionality of the project.