chore: remove console.log statement from server/** directory
This commit removes the console.log statement that was added to the code in
the server/** directory to test whether the GitHub workflows were working
properly. Since the test has been successful, the statement is no longer needed
and has been removed to keep the code clean and free of unnecessary clutter.