feat: add new GitHub workflow action to deploy Kubernetes manifests
This commit adds a new action file to the workflows directory to automate the
deployment of Kubernetes manifests for the project. The action will run on
every push to the main branch and will use the manifests file to deploy the
application to the Kubernetes cluster. This change will streamline the
deployment process and make it easier to manage the project's infrastructure.