ci: update Kubernetes deployment file for testing GitHub action
This commit updates the Kubernetes deployment file to test the new GitHub
workflow action that was added to automate the deployment process. The file
has been updated to reflect changes in the project's infrastructure and to
ensure that the deployment process is reliable and secure. This change will
help to identify and fix any issues with the deployment process early on and
improve the overall stability of the project.